added backup and import functionality.
Backup is a download function which, when button is clicked a text file copy of the entire inventory as it is laid out in the JS objects inside the main array.

The file downloads to the default download folder so the user will have to copy/paste it to a specific folder where it can be retrieved at time of import.  I hope to expand on the function so that a 'save-as' dialogue box appears when the backup button is clicked, giving the user the option to save wherever they wish; at which time they can also choose to create a new folder for the text file.

One issue is with the import where, for some reason, the text file for the inventory, which is a stringified JSON object, cannot be converted due to some 'non-whitespace' error at time of the attempted conversion.  I tested the text file by copying and pasting into JSON.lint online, and the test came back valid.

JSON lint formatted the file for readability, and so I copy/pasted the text from lint and pasted it over the original text file.  The file was then accepted as valid JSON when the import function was used on the altered file.  So, even though the text file contains valid JSON format, initially it cannot be used.

Currently, two things are backed up, the inventory as it is laid out with nested objects and arrays, and an array containing objects representing each item in the inventory.  The latter imports with no problems. So currently I have to find out why the first file is not formatting correctly.  Once that is done I can complete the import process by writing a function that allows user locate the backup, perform an import and use it as a fresh inventory if user has opens the app on a secondary device.

In a future iteration I'm considering saving my inventory to a database, such as MongoDB and using REACT as a front-end to view and modify the inventory; as I'm just learning MongoDB.  Whether this newer version would be usable by people other than me is another story, because I assume they would have to set up their own database, and the use of local storage and the text backup requires only a small amount of effort on the part of users, whereas the mongoDB database is much more involved.

I've also considered using IndexedDB as a means of storage.  All of these ideas stemmed from my own fears when I heard that both localStorage and IndexedDB are cleared by some browsers after a period of time; and if users have a large amount of items in their inventory, all recorded in the app, they might lose all of that information, hence why I set up the simple backup system.