backup text file conversion to JSON issue solved:
Instead of using JSON.stringify(container) as an argument to the new Blob used to create a text file from the inventory, if did it outside of the function and assigned it a variable, nesting the method inside an array, and then used the variable as an argument to the new Blob.  This solved the issue and now, on importing the text file, it is successfully converted back to a javascript object using JSON.parse().