changed wording slightly on BackupModal component; instead of using 'Backup filename:' to describe the name of the downloaded backup file, I've changed that to 'Backup filename:'
Also, checked to see that backup only occurs when inventory is not empty; otherwise user will receive an alert indicating that there is no inventory.

I tested the import function and it is functional with no issues so far.

for the next iteration I'm looking at ways to save the database, either through indexedDB or MongoDB.  The ultimate goal would be for the user to have their database on the cloud and for there to be an authentication system so they can access the inventory from any device; then there  would be just one inventory, meaning there would be no need to keep separate inventories up to date, and also, in the case of device failure, the database would not be corrupted.

Other ideas:
Have the backups saved remotely in a 'cloud'.
If the inventory is modified, give the user the option do a backup; perhaps have an extra button appear on the start page so each time the user opens the app,it serves as a reminder that a backup hasn't been performed since the last inventory edit.